integrated prompt logic with main page
